Fixed the issues with my project.
I have added all of the photos used for this project. I tried moving the whole folder of images and they were all removed from it. I am not sure if it is an error that is on my end or if GitHub does that naturally. If it is a problem, please let me know!